bug 1480661 use single precision for operating on delay values r=padenot
These are derived from single precision AudioParam values
and so there is little to gain from using double.

double has the disadvantage that values intended to be integer
frame counts are less likely to be integer. e.g.

(rr) print ((float)(128.0/48000.0))*48000.f
$91 = 128
(rr) print ((float)(128.0/48000.0))*48000.0
$92 = 127.99999862909317
(rr) print ((float)(128.0/44100.0))*44100.f
$93 = 128
(rr) print ((float)(128.0/44100.0))*44100.0
$94 = 127.99999765120447
(rr) print ((float)(768.0/48000.0))*48000.f
$95 = 768.000061
(rr) print ((float)(768.0/48000.0))*48000.0
$96 = 768.0000364780426
(rr) print ((float)(768.0/44100.0))*44100.f
$97 = 768
(rr) print ((float)(768.0/44100.0))*44100.0
$98 = 768.00002697855234

MozReview-Commit-ID: FlmZJeBjUYY

UltraBlame original commit: b3f01b9ce2bf9c4cf05bf4bca16330560439db0b

+++ dom/media/webaudio/DelayBuffer.h
@@ -19,9 +19,9 @@ class DelayBuffer final
   typedef dom::ChannelInterpretation ChannelInterpretation;
 
 public:
-  explicit DelayBuffer(double aMaxDelayTicks)
+  explicit DelayBuffer(float aMaxDelayTicks)
     
-    : mMaxDelayTicks(ceil(aMaxDelayTicks))
+    : mMaxDelayTicks(std::ceil(aMaxDelayTicks))
     , mCurrentChunk(0)
     
 #ifdef DEBUG
@@ -40,19 +40,19 @@ public:
 
   
   
-  void Read(const double aPerFrameDelays[WEBAUDIO_BLOCK_SIZE],
+  void Read(const float aPerFrameDelays[WEBAUDIO_BLOCK_SIZE],
             AudioBlock* aOutputChunk,
             ChannelInterpretation aChannelInterpretation);
   
   
-  void Read(double aDelayTicks, AudioBlock* aOutputChunk,
+  void Read(float aDelayTicks, AudioBlock* aOutputChunk,
             ChannelInterpretation aChannelInterpretation);
 
   
   
   
   
-  void ReadChannel(const double aPerFrameDelays[WEBAUDIO_BLOCK_SIZE],
+  void ReadChannel(const float aPerFrameDelays[WEBAUDIO_BLOCK_SIZE],
                    AudioBlock* aOutputChunk, uint32_t aChannel,
                    ChannelInterpretation aChannelInterpretation);
 
@@ -75,7 +75,7 @@ public:
   size_t SizeOfExcludingThis(MallocSizeOf aMallocSizeOf) const;
 
 private:
-  void ReadChannels(const double aPerFrameDelays[WEBAUDIO_BLOCK_SIZE],
+  void ReadChannels(const float aPerFrameDelays[WEBAUDIO_BLOCK_SIZE],
                     AudioBlock* aOutputChunk,
                     uint32_t aFirstChannel, uint32_t aNumChannelsToRead,
                     ChannelInterpretation aChannelInterpretation);
